Restructured stream classes to make space for custom scheduling logic Fixed bugs related to custom scheduling when operators produce more than 1 tuple at a time

+++ src/main/java/stream/BoundedStream.java
@@ -93,12 +93,12 @@ public class BoundedStream<T extends Tuple> implements Stream<T> {
   }
 
   @Override
-  public long size() {
+  public int size() {
     return stream.size();
   }
 
   @Override
-  public long remainingCapacity() {
+  public int remainingCapacity() {
     return stream.remainingCapacity();
   }
